+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We use [semantic versioning](http://semver.org/):
 
 # Next version
 - [feature] _agent_: Experimental support for Java 25 class files
+- [fix] Erroneous log messages were printed about unmatched probes
 
 # 35.0.3
 - [fix] _teamscale-gradle-plugin_: `IllegalArgumentException`: Can't execute request without discovering it first